Fix cZone background storage path to match the real deploy layout
Matt confirmed the actual production layout: /var/www/cartoon-reorbit-images
is a sibling directory to /var/www/cartoon-reorbit (the repo checkout),
kept outside the repo for size. Replaced the three background filesystem
endpoints' per-file __dirname-based ".. count to reach the root" math
(which computed a directory relative to the repo root, not its parent,
and was verified broken against a real production build regardless) with
a shared helper that goes one directory up from process.cwd() -- reliably
the repo root in both `nuxt dev` and production, since PM2 launches
.output/server/index.mjs with the repo root as cwd per the relative
script path in ecosystem.config.cjs.

Verified against a simulated /var/www/cartoon-reorbit + sibling
/var/www/cartoon-reorbit-images layout that the computed upload
directory lands in the right place.

The layoutData propagation logic (updating CZone.background,
CZone.layoutData.zones[].background, and
CZoneSearchPrize.conditionBackgrounds when a background is edited or
deleted) was already correct -- verified by running the exact
transaction SQL from both endpoints against a live Postgres 16 seeded
with a CZone referencing a background at both denormalized locations;
only the matching reference updated, unrelated references were left
untouched. Left unchanged.
