+++ server/api/admin/backgrounds/[id].delete.js
@@ -1,24 +1,10 @@
 import { defineEventHandler, getRequestHeader, createError } from 'h3'
-import { join, dirname } from 'node:path'
-import { fileURLToPath } from 'node:url'
 import { unlink } from 'node:fs/promises'
 import { prisma as db } from '@/server/prisma'
 import { logAdminChange } from '@/server/utils/adminChangeLog'
 import { redis } from '@/server/utils/redis'
 import { clearSearchesCache } from '@/server/api/czone/[username]/searches.get'
-
-const __dirname = dirname(fileURLToPath(import.meta.url))
-// In production this file is at server/api/admin/backgrounds/ — 4 levels up to project root.
-// In dev, process.cwd() is always the project root.
-const baseDir = process.env.NODE_ENV === 'production'
-  ? join(__dirname, '..', '..', '..', '..')
-  : process.cwd()
-
-function fsPathFromFilename(filename) {
-  return process.env.NODE_ENV === 'production'
-    ? join(baseDir, 'cartoon-reorbit-images', 'backgrounds', filename)
-    : join(baseDir, 'public', 'backgrounds', filename)
-}
+import { backgroundFsPath } from '@/server/utils/backgroundStorage'
 
 export default defineEventHandler(async (event) => {
   const cookie = getRequestHeader(event, 'cookie') || ''
@@ -117,7 +103,7 @@ export default defineEventHandler(async (event) => {
   }
 
   // DB committed — delete image file from disk (non-fatal if missing)
-  try { await unlink(fsPathFromFilename(bg.filename)) } catch {}
+  try { await unlink(backgroundFsPath(bg.filename)) } catch {}
 
   // Invalidate Redis cache for any cZone searches that had this background as a condition
   for (const searchId of affectedSearchIds) {
