+++ server/api/admin/backgrounds/[id]/replace-image.post.js
@@ -1,31 +1,18 @@
 import { defineEventHandler, readMultipartFormData, getRequestHeader, createError } from 'h3'
 import { mkdir, writeFile, unlink } from 'node:fs/promises'
-import { join, dirname } from 'node:path'
-import { fileURLToPath } from 'node:url'
+import { join } from 'node:path'
 import { prisma as db } from '@/server/prisma'
 import { imageSize } from 'image-size'
 import { logAdminChange } from '@/server/utils/adminChangeLog'
 import { redis } from '@/server/utils/redis'
 import { clearSearchesCache } from '@/server/api/czone/[username]/searches.get'
-
-const __dirname = dirname(fileURLToPath(import.meta.url))
-// In production this file is at server/api/admin/backgrounds/[id]/ — 5 levels up to project root.
-// In dev, process.cwd() is always the project root.
-const baseDir = process.env.NODE_ENV === 'production'
-  ? join(__dirname, '..', '..', '..', '..', '..')
-  : process.cwd()
+import { backgroundUploadDir, backgroundFsPath, backgroundPublicPath } from '@/server/utils/backgroundStorage'
 
 const ALLOWED_MIMES = ['image/png', 'image/jpeg', 'image/gif']
 const ALLOWED_SIZES = [[510, 344], [512, 346], [800, 600]]
 
 function sanitize(name = '') {
   return name.replace(/[^A-Za-z0-9._-]/g, '') || 'image'
-}
-
-function fsPathFromFilename(filename) {
-  return process.env.NODE_ENV === 'production'
-    ? join(baseDir, 'cartoon-reorbit-images', 'backgrounds', filename)
-    : join(baseDir, 'public', 'backgrounds', filename)
 }
 
 export default defineEventHandler(async (event) => {
@@ -63,9 +50,7 @@ export default defineEventHandler(async (event) => {
     throw createError({ statusCode: 400, statusMessage: `Image must be exactly ${supported}.` })
   }
 
-  const uploadDir = process.env.NODE_ENV === 'production'
-    ? join(baseDir, 'cartoon-reorbit-images', 'backgrounds')
-    : join(baseDir, 'public', 'backgrounds')
+  const uploadDir = backgroundUploadDir()
 
   await mkdir(uploadDir, { recursive: true })
 
@@ -74,9 +59,7 @@ export default defineEventHandler(async (event) => {
   const outPath = join(uploadDir, newFilename)
   await writeFile(outPath, imagePart.data)
 
-  const newImagePath = process.env.NODE_ENV === 'production'
-    ? `/images/backgrounds/${newFilename}`
-    : `/backgrounds/${newFilename}`
+  const newImagePath = backgroundPublicPath(newFilename)
 
   const oldFilename = bg.filename
   const oldImagePath = bg.imagePath
@@ -159,7 +142,7 @@ export default defineEventHandler(async (event) => {
   }
 
   // DB committed — now safe to remove the old file
-  try { await unlink(fsPathFromFilename(oldFilename)) } catch {}
+  try { await unlink(backgroundFsPath(oldFilename)) } catch {}
 
   // Invalidate Redis cache for any affected cZone searches
   for (const searchId of affectedSearchIds) {
